% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_functions.R
\name{add_date_info}
\alias{add_date_info}
\title{add date info to data frame}
\usage{
add_date_info(df, date_df)
}
\arguments{
\item{df}{data frame of quadrat data. Must contain columns "quadrat" and "year"}

\item{date_df}{data frame of date information to be merged with df, containing columns "quadrat" and "year"}
}
\description{
add date info to data frame
}
