% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_functions.R
\name{add_species_info}
\alias{add_species_info}
\title{add species info to data frame}
\usage{
add_species_info(df, column_name, species_df)
}
\arguments{
\item{df}{data frame of quadrat data containing a column for USDA species code}

\item{column_name}{name of column from df containing USDA species code}

\item{species_df}{data frame of species information to be merged with df, containing a column USDA_code}
}
\description{
add species info to data frame
}
