% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrat_qc_functions.R
\name{check_species_codes}
\alias{check_species_codes}
\title{Check species codes}
\usage{
check_species_codes(df, column_name, specieslist)
}
\arguments{
\item{df}{Data frame of quadrat data to validate}

\item{column_name}{Name of column in df containing species codes}

\item{specieslist}{Vector of valid species codes}
}
\value{
Rows from df containing invalid species codes, plus additional columns indicating error_type and error_value.
       If there are no errors, this will return an empty data frame.
}
\description{
Check a dataframe of quadrat data for valid plant species codes. Will flag species codes that are the incorrect case (upper/lower), or NA and missing values
}
\examples{
check_species_codes(dataframe, column_name = 'Species', specieslist = sp_list$USDA_codes)

}
