% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_functions.R
\name{create_count_df}
\alias{create_count_df}
\title{create count data frame}
\usage{
create_count_df(df, quad_info, desired_columns)
}
\arguments{
\item{df}{data frame of point data created by readOGR}

\item{quad_info}{information about quadrat name and year gleaned from layer file name}

\item{desired_columns}{vector of column names that should be extracted from the attribute table}
}
\description{
creates data frame of count data from point data layer from gdb
}
