% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_functions.R
\name{create_cover_df}
\alias{create_cover_df}
\title{create cover data frame}
\usage{
create_cover_df(df, quad_info, sp_list, dates_df)
}
\arguments{
\item{df}{data frame of polygon data created by readOGR}

\item{quad_info}{information about quadrat name and year gleaned from layer file name}

\item{sp_list}{data frame of species information}

\item{dates_df}{data frame of quadrat sampling dates}
}
\description{
creates data frame of cover data from polygon data layer from gdb
}
