% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_functions.R
\name{error_check_point}
\alias{error_check_point}
\title{error check point data frame}
\usage{
error_check_point(df, sp_list)
}
\arguments{
\item{df}{data frame of point data created by readOGR}

\item{sp_list}{data frame of species information}
}
\description{
checks point layer of quadrat data from a gdb file for errors
}
