% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_functions.R
\name{error_check_polygon}
\alias{error_check_polygon}
\title{error check polygon data frame}
\usage{
error_check_polygon(df, sp_list)
}
\arguments{
\item{df}{data frame of polygon data created by readOGR}

\item{sp_list}{data frame of species information}
}
\description{
checks polygon layer of quadrat data from a gdb file for errors
}
