% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrat_qc_functions.R
\name{find_duplicate_rows}
\alias{find_duplicate_rows}
\title{Find duplicate rows}
\usage{
find_duplicate_rows(df, columns)
}
\arguments{
\item{df}{Data frame of quadrat data}

\item{columns}{Vector of column names that determine unique key. (e.g. columns = c('quadrat', 'date', 'species'))}
}
\value{
Rows from df that are duplicates according to the given key, plus an additional column indicating error_type.
        If there are no errors, this will return an empty data frame.
}
\description{
Check a data frame of quadrat data for duplicate data (possibly data entry error). User supplies
  list of "key" column names that should define a unique entry (e.g. quadrat, date, species should be associated with
  a single count value). This function ignores rows with NA or missing values in the key columns.
}
\examples{
find_duplicate_rows(df, columns = c('quadrat', 'year', 'month', 'USDA_code'))

}
