% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_functions.R
\name{create_count_df}
\alias{create_count_df}
\title{create count data frame}
\usage{
create_count_df(df, quad_info, sp_list, dates_df)
}
\arguments{
\item{df}{data frame of point data created by readOGR}

\item{quad_info}{information about quadrat name and year gleaned from layer file name}

\item{sp_list}{data frame of species information}

\item{dates_df}{data frame of quadrat sampling dates}
}
\description{
creates data frame of count data from point data layer from gdb
}
